fix spectator event to fire for all applicable game states
